[#10]refactor: Auth 관련 패키지 User-> Auth 패키지로 이동
