+++ src/main/java/org/example/mollyapi/user/controller/SignUpController.java
@@ -1,4 +1,4 @@
-package org.example.mollyapi.user;
+package org.example.mollyapi.user.controller;
 
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
